Extend Redmine library to support non GET method. Sample code for post an issue is implemented in Post.hs.
- JSON related code originaly in Get.hs are moved to JSON.hs.
- ToJSON code are added in JSON.hs
- Code to request for Redmine server are moved to Rest.hs from Get.hs to share between Get.hs and Post.hs
 - runQuery or other code are extended to support POST or other method and to accept data to be sent to Redmine Server.
- code usable for user are moved to Utils.hs
