+++ Redmine.cabal
@@ -1,5 +1,5 @@
 name: Redmine
-version: 0.0.3
+version: 0.0.5
 cabal-version: >=1.10
 build-type: Simple
 license: MIT
@@ -19,7 +19,7 @@ source-repository head
 Library
     default-language: Haskell2010
     build-depends: base >=2 && <5, old-locale -any,
-                   time -any, old-time -any, bytestring -any, http-conduit >= 2.1.0,
+                   time -any, old-time -any, bytestring -any, http-conduit >= 2.1.0, http-types, 
                    aeson -any, HTTP -any, network -any, text,
                    transformers, containers, http-client-tls,
                    resourcet, connection, MissingH
